# src/main/resources/public/docs/swagger.json
{"openapi":"3.0.1","info":{"title":"SuperAPI","version":"1.0.0"},"servers":[{"url":"http://localhost:8080/"}],"paths":{"/Login":{"post":{"tags":["Login"],"summary":"Login","operationId":"login","requestBody":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"default":{"description":"default response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/users":{"get":{"tags":["Users API v1.0.0-SHOWTIME"],"summary":"Find a user","operationId":"findUserById","parameters":[{"name":"Authorization","in":"header","schema":{"type":"string"}},{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"default":{"description":"default response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"post":{"tags":["Users API v1.0.0-SHOWTIME"],"summary":"Creates a new user","operationId":"createUser","requestBody":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/NewUserDTO"}}}},"responses":{"default":{"description":"default response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"components":{"schemas":{"PhoneDTO":{"type":"object","properties":{"number":{"type":"string"},"ddd":{"type":"string"}}},"UserDTO":{"type":"object","properties":{"last_login":{"type":"string","format":"date-time","writeOnly":true},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"email":{"type":"string"},"phones":{"type":"array","items":{"$ref":"#/components/schemas/PhoneDTO"}},"created":{"type":"string","format":"date-time"},"modified":{"type":"string","format":"date-time"},"token":{"type":"string"},"lastLogin":{"type":"string","format":"date-time"}}},"LoginDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"NewUserDTO":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"phones":{"type":"array","items":{"$ref":"#/components/schemas/PhoneDTO"}},"password":{"type":"string"}}}}}}